// src/types/Page.ts
export type Background = Custom | Standard | Swoosh | Container;
// ...
export interface Image {
  kind?: "image";
  title?: string | null;
  src: string;
  width: number | string;
  height: number | string;
  alt?: string;
  hint?: string;
  is_decorative?: boolean;
  typeOptions?: {
    [k: string]: unknown;
  };
  sizes?: Size[];
  image_background_circle_color?: string;
  image_background_gradient_color?: string;
  image_alignment?: string;
  file_url?: string;
}
export interface Size {
  src?: string;
  width?: number;
  height?: number;
  [k: string]: unknown;
}
// ...
export interface Cta {
  /**
   * The label of the button.
   */
  label?: string;
  ariaLabel?: string;
  url?: string;
  /**
   * Icon identifier
   */
  icon?: string;
  variant?: "primary" | "secondary" | "tertiary" | "text" | "trial";
  target?: "_self" | "_parent" | "_top" | "_blank";
  id?: string | number;
  source?: "standard-link" | "modal-link" | "custom-action-link";
  custom_action_options?: {
    event_code?: string;
    workflow_uri?: string;
    name_capture_campaign_id?: string;
    action_template?: "everybody-demo-org" | "event-sign-up" | "offer";
    [k: string]: unknown;
  };
  linkAttributes?: {
    key?: string;
    value?: string;
    [k: string]: unknown;
  }[];
  modal_variant?: string;
  [k: string]: unknown;
}
// ...
export interface Custom {
  kind: "custom_background";
  blade_background_color?: string;
  blade_custom_image_desktop: Image;
  blade_custom_image_mobile: Image;
  blade_background_gradient_direction?: string;
}
export interface Standard {
  kind: "standard_background";
  blade_background_color?: string;
  blade_background_image?: Image;
  blade_background_shape?: string;
  blade_background_gradient_direction?: string;
}
export interface Swoosh {
  kind: "swoosh_background";
  blade_background_color: string;
}
export interface Container {
  kind: "container_background";
  blade_background_color?: string;
  blade_background_image?: Image;
  blade_background_image_position?: string;
  blade_background_shape?: string;
  blade_background_gradient_direction?: string;
}
// ...
export interface CardQuote {
  id: string;
  kind: "card_quote";
  align_image?: "left" | "right";
  quote_image?: {
    headshot?: Image;
    logo?: Image;
    product?: Image;
    [k: string]: unknown;
  };
  quote_text: string;
  name: string;
  role: string;
  company_name: string;
  link_1?: Cta;
  card_background_type?: Background;
}